feat(models): move bulk_load_checks_metadata inside class
